requirements.txt
fastapi==0.115.4
fastapi-cli==0.0.5
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pymongo==4.10.1